stem-separator/requirements.txt
# Core ML/Audio
torch>=2.0.0
torchaudio>=2.0.0
demucs>=4.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
audioread>=3.0.0

# Note Detection & MIDI
basic-pitch>=0.2.0
pretty_midi>=0.2.10
midiutil>=1.2.1

# Analysis
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Claude API
anthropic>=0.18.0

# CLI & UI
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
streamlit>=1.30.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# File Watching
watchdog>=3.0.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.0
